+++ client/src/components/Container/style.css
@@ -5,6 +5,5 @@
   padding: 10px;
   padding-bottom: 10px;
   border-radius: 20px;
-  overflow: scroll;
   min-height: 65vh;
 }
