+++ client/src/pages/Home/index.js
@@ -5,7 +5,7 @@ import HomeNav from "../../components/HomeNav";
 import { Col, Row } from "../../components/Grid";
 import Container from "../../components/Container";
 import Teachers from "./images/teachers.jpg";
-import UI from "./images/UI.jpg";
+
 function Home() {
   return (
     <div>
