--- client/src/pages/Students/style.css
@@ -1,1 +1,0 @@
-
